test: reorganize misplaced test files from src/ to test/ directory
- Consolidated duplicate json.test.ts files by merging summarizeEvaluateResultForLogging tests from src/util/json.test.ts into existing test/util/json.test.ts
- Moved src/server/routes/eval.test.ts to test/server/routes-eval.test.ts and updated import paths
- Removed duplicate test files to follow project convention of keeping core logic tests in test/ directory
- All moved tests continue to pass and maintain existing functionality
